groupware: refactoring: pass object type instead of namespaces
 * make the JMAP internal API a bit more future-proof by passing
   ObjectType objects instead of the JMAP namespaces

 * remove the new attempt to contain operations even further using the
   Factory objects

 * move CalendarEvent operations to its own file, like everything else

 * fix email tests

 * ignore WS error when closing an already closed connection

+++ pkg/jmap/integration_calendar_test.go
@@ -126,7 +126,7 @@ func TestEvents(t *testing.T) {
 			results := m[accountId]
 			require.Equal(len(results.Results), int(page))
 			require.Equal(limit, results.Limit)
-			require.Equal(position, results.Position)
+			require.Equal(uint(position), results.Position)
 			require.Equal(true, results.CanCalculateChanges)
 			require.NotNil(results.Total)
 			require.Equal(count, *results.Total)
